% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/util.R
\name{AddTrans}
\alias{AddTrans}
\title{Add transpancy to color}
\usage{
AddTrans(col, trans)
}
\arguments{
\item{col}{A set of colors}

\item{trans}{A vector or vector of the same length as col,
0 being fully visable and 1 being fully transpant}
}
\description{
Add transpancy to color
}
\author{
Jiarui Ding
}

