% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{AML28}
\alias{AML28}
\title{The variant allele frequencies of somatic single nucleotide variants
from whole genome sequencing of a AML patient at two time points}
\format{A list with two fields
\itemize{
  \item x. A matrix of allele frequencies
  \item y. Clustering labels from the original publication
}}
\usage{
AML28
}
\description{
The variant allele frequencies of somatic single nucleotide variants
from whole genome sequencing of a AML patient at two time points
}
\references{
Miller, C. A., White, B. S., Dees, N. D., Griffith, M., Welch, J. S., Griffith, O. L.,
Vij, R., Tomasson, M. H., Graubert, T. A., Walter, M. J., et al. (2014).
SciClone: Inferring clonal architecture and tracking the spatial and temporal patterns
of tumor evolution. PLoS Comput. Biol., 10(8).
}
\keyword{datasets}

