% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/util.R
\name{AssignLabelColor}
\alias{AssignLabelColor}
\title{Assign colors to labels}
\usage{
AssignLabelColor(col, label, uniq.label)
}
\arguments{
\item{col}{A set of colors}

\item{label}{Cluster labels}

\item{uniq.label}{The initial labels}
}
\description{
Assign colors to labels
}
\author{
Jiarui Ding
}

