% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/util.R
\name{ComputeNMI}
\alias{ComputeNMI}
\title{Calcular the normalized mutual information}
\usage{
ComputeNMI(ground.truth, cluster)
}
\arguments{
\item{ground.truth}{The cluster ground truth}

\item{cluster}{The cluster labels}
}
\description{
Calcular the normalized mutual information
}

