% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cytof.benchmark.h2}
\alias{cytof.benchmark.h2}
\title{Single-cell CyTOF data of bone marrow mononuclear cells from a healthy individual H2}
\format{A list with two fields
\itemize{
  \item x. A cell-protein markers matrix of protein expression values
  \item metadata. A list with two fields: 1) the cell type of each cell and (2) the t-SNE coordinates
}}
\usage{
cytof.benchmark.h2
}
\description{
Single-cell CyTOF data of bone marrow mononuclear cells from a healthy individual H2
}
\references{
Levine, J. H., Simonds, E. F., Bendall, S. C., Davis, K. L., El-ad, D. A., Tadmor,
M. D., Litvin, O., Fienberg, H. G., Jager, A., Zunder, E. R., et al. (2015).
Data- driven phenotypic dissection of aml reveals progenitor-like cells
that correlate with prognosis. Cell, 162(1), 184-197.
}
\keyword{datasets}

