% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cytof.benchmark.one}
\alias{cytof.benchmark.one}
\title{Single-cell CyTOF data of bone marrow mononuclear cells from a healthy individual}
\format{A list with two fields
\itemize{
  \item x. A cell-protein markers matrix of protein expression values
  \item metadata. A list with two fields: 1) the cell types of each cell and (2) the t-SNE coordinates
}}
\usage{
cytof.benchmark.one
}
\description{
Single-cell CyTOF data of bone marrow mononuclear cells from a healthy individual
}
\references{
Bendall, S. C., Simonds, E. F., Qiu, P., El-ad, D. A., Krutzik, P. O., Finck, R., Bruggner,
R. V., Melamed, R., Trejo, A., Ornatsky, O. I., et al. (2011).
Single-cell mass cytometry of differential immune and drug responses
across a human hematopoietic continuum. Science, 332(6030), 687-696.
}
\keyword{datasets}

