% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{distinct.col}
\alias{distinct.col}
\title{A vector of 298 visually distinctive colours}
\format{Each element of the vector is a colour hex code}
\source{
\url{http://graphicdesign.stackexchange.com/revisions/3815/8} and other places
}
\usage{
distinct.col
}
\description{
A vector of 298 visually distinctive colours
}
\keyword{datasets}

