% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{MEL5}
\alias{MEL5}
\title{The variant allele frequencies of somatic single nucleotide variants
from whole genome sequencing of the lung/pancreatic metastasis of a melanoma patient}
\format{A list with one field
\itemize{
  \item x. A matrix of allele frequencies
}}
\usage{
MEL5
}
\description{
The variant allele frequencies of somatic single nucleotide variants
from whole genome sequencing of the lung/pancreatic metastasis of a melanoma patient
}
\references{
Ding, L., Kim, M., Kanchi, K. L., Dees, N. D., Lu, C., Griffith, M., Fenstermacher,
 D., Sung, H., Miller, C. A., Goetz, B., et al. (2014). Clonal architectures and driver
 mutations in metastatic melanomas. PLoS ONE, 9(11).
}
\keyword{datasets}

