% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{PMF}
\alias{PMF}
\title{The variant allele frequencies of somatic single nucleotide variants
from whole genome sequencing of a primary myelofibrosis patient at three time points}
\format{A list with three fields
\itemize{
  \item x. A matrix of allele frequencies
  \item y. Clustering labels from the original publication
  \item metadata. A list with one field: the gene names of each mutation
}}
\usage{
PMF
}
\description{
The variant allele frequencies of somatic single nucleotide variants
from whole genome sequencing of a primary myelofibrosis patient at three time points
}
\references{
Engle, E., Fisher, D., Miller, C., McLellan, M., Fulton, R., Moore, D., Wilson, R., Ley,T., and Oh, S. (2015).
Clonal evolution revealed by whole genome sequencing in a case of primary myelofibrosis transformed to
secondary acute myeloid leukemia. Leukemia, 29(4), 869-876.
}
\keyword{datasets}

