% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{single.cell.mrna.florens}
\alias{single.cell.mrna.florens}
\title{Single cell gene expression data from 223 stem cells
from the subentricular zone of eight-week-old male mice}
\format{A list with two fields
\itemize{
  \item x. A cell-gene matrix of gene expression values
  \item metadata. A list with only one field: the cell type of each cell
}}
\usage{
single.cell.mrna.florens
}
\description{
Single cell gene expression data from 223 stem cells
from the subentricular zone of eight-week-old male mice
}
\references{
Llorens-Bobadilla, E., Zhao, S., Baser, A., Saiz-Castro, G., Zwadlo, K., and Martin-Villalba, A. (2015).
Single-cell transcriptomics reveals a population of dormant neural stem cells
that become activated upon brain injury. Cell Stem Cell. 10.1016/j.stem.2015.07.002.
}
\keyword{datasets}

