% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{single.cell.mrna.pollen}
\alias{single.cell.mrna.pollen}
\title{Single cell gene expression data from 11 populations}
\format{A list with two fields
\itemize{
  \item x. A cell-gene matrix of gene expression values
  \item metadata. A list with one field: the cell type of each cell
}}
\usage{
single.cell.mrna.pollen
}
\description{
Single cell gene expression data from 11 populations
}
\references{
Pollen, A. A., Nowakowski, T. J., Shuga, J., Wang, X., Leyrat, A. A., Lui, J. H., Li, N.,
Szpankowski, L., Fowler, B., Chen, P., et al. (2014).
Low-coverage single-cell mrna sequencing reveals cellular heterogeneity
and activated signaling pathways in developing cerebral cortex.
Nat. Biotechnol., 32(10), 1053-1058.
}
\keyword{datasets}

